standardize connector layout for postgres + mysql
