+++ connectors/mysql/internal/connection/connection.go
@@ -13,68 +13,51 @@ import (
 	"github.com/galaxy-io/filament/connectors/internal/dbconfig"
 )
 
-const defaultPort = 3306
-
-// Fields returns the configuration fields for a MySQL connection.
-func Fields() []filament.ConfigField {
-	fields := dbconfig.VisibleWhen(dbconfig.MethodFields)
-	return []filament.ConfigField{
-		dbconfig.MethodConfigField(),
-		dbconfig.DSNConfigField("MySQL URL or driver DSN"),
-		{Name: "host", Type: filament.FieldString, Required: true, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "MySQL server hostname"},
-		{Name: "port", Type: filament.FieldInt, Default: defaultPort, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "MySQL server port"},
-		{Name: "username", Type: filament.FieldString, Required: true, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "MySQL username"},
-		{Name: "password", Type: filament.FieldSecret, Required: true, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "MySQL password"},
-		{Name: "database_name", Type: filament.FieldString, Required: true, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "Database to connect to"},
-		{Name: "tls_mode", Type: filament.FieldEnum, Default: "preferred", Enum: []filament.EnumOption{
-			{Value: "false", Label: "Disable"},
-			{Value: "preferred", Label: "Prefer"},
-			{Value: "true", Label: "Require"},
-			{Value: "skip-verify", Label: "Require, skip verification"},
-		}, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "MySQL TLS mode"},
-	}
+// Resolved contains the native MySQL driver configuration.
+type Resolved struct {
+	DriverConfig *mysql.Config
 }
 
 // Resolve translates connector configuration into a MySQL driver config.
-func Resolve(cfg filament.Config) (*mysql.Config, error) {
+func Resolve(cfg filament.Config) (Resolved, error) {
 	method, err := dbconfig.Method(cfg)
 	if err != nil {
-		return nil, err
+		return Resolved{}, err
 	}
 	if method == dbconfig.MethodURL {
 		dsn := cfg.Secret(dbconfig.DSNField)
 		if strings.TrimSpace(dsn) == "" {
-			return nil, fmt.Errorf("dsn is required when connection_method is %q", dbconfig.MethodURL)
+			return Resolved{}, fmt.Errorf("dsn is required when connection_method is %q", dbconfig.MethodURL)
 		}
 		parsed, err := parseDSN(dsn)
 		if err != nil {
-			return nil, fmt.Errorf("parse dsn: %w", err)
+			return Resolved{}, fmt.Errorf("parse dsn: %w", err)
 		}
-		return parsed, nil
+		return Resolved{DriverConfig: parsed}, nil
 	}
 
 	host := strings.TrimSpace(cfg.String("host"))
 	if host == "" {
-		return nil, fmt.Errorf("host is required when connection_method is %q", dbconfig.MethodFields)
+		return Resolved{}, fmt.Errorf("host is required when connection_method is %q", dbconfig.MethodFields)
 	}
 	port := cfg.Int("port")
 	if port == 0 {
 		port = defaultPort
 	}
 	if port < 1 || port > 65535 {
-		return nil, fmt.Errorf("port must be between 1 and 65535")
+		return Resolved{}, fmt.Errorf("port must be between 1 and 65535")
 	}
 	username := cfg.String("username")
 	if username == "" {
-		return nil, fmt.Errorf("username is required when connection_method is %q", dbconfig.MethodFields)
+		return Resolved{}, fmt.Errorf("username is required when connection_method is %q", dbconfig.MethodFields)
 	}
 	password := cfg.Secret("password")
 	if password == "" {
-		return nil, fmt.Errorf("password is required when connection_method is %q", dbconfig.MethodFields)
+		return Resolved{}, fmt.Errorf("password is required when connection_method is %q", dbconfig.MethodFields)
 	}
 	database := cfg.String("database_name")
 	if database == "" {
-		return nil, fmt.Errorf("database_name is required when connection_method is %q", dbconfig.MethodFields)
+		return Resolved{}, fmt.Errorf("database_name is required when connection_method is %q", dbconfig.MethodFields)
 	}
 	tlsMode := cfg.String("tls_mode")
 	if tlsMode == "" {
@@ -83,7 +66,7 @@ func Resolve(cfg filament.Config) (*mysql.Config, error) {
 	switch tlsMode {
 	case "false", "preferred", "true", "skip-verify":
 	default:
-		return nil, fmt.Errorf("unsupported tls_mode %q", tlsMode)
+		return Resolved{}, fmt.Errorf("unsupported tls_mode %q", tlsMode)
 	}
 	host = strings.TrimPrefix(strings.TrimSuffix(host, "]"), "[")
 	resolved := mysql.NewConfig()
@@ -93,7 +76,11 @@ func Resolve(cfg filament.Config) (*mysql.Config, error) {
 	resolved.Addr = net.JoinHostPort(host, strconv.Itoa(port))
 	resolved.DBName = database
 	resolved.TLSConfig = tlsMode
-	return mysql.ParseDSN(resolved.FormatDSN())
+	driverConfig, err := mysql.ParseDSN(resolved.FormatDSN())
+	if err != nil {
+		return Resolved{}, err
+	}
+	return Resolved{DriverConfig: driverConfig}, nil
 }
 
 // parseDSN accepts both the go-sql-driver native DSN and the conventional
