+++ connectors/postgres/internal/connection/connection.go
@@ -13,34 +13,10 @@ import (
 	"github.com/galaxy-io/filament/connectors/internal/dbconfig"
 )
 
-const defaultPort = 5432
-
 // Resolved contains the PostgreSQL connection string and parsed pool config.
 type Resolved struct {
-	DSN  string
-	Pool *pgxpool.Config
-}
-
-// Fields returns the configuration fields for a PostgreSQL connection.
-func Fields() []filament.ConfigField {
-	fields := dbconfig.VisibleWhen(dbconfig.MethodFields)
-	return []filament.ConfigField{
-		dbconfig.MethodConfigField(),
-		dbconfig.DSNConfigField("PostgreSQL connection URL"),
-		{Name: "host", Type: filament.FieldString, Required: true, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "PostgreSQL server hostname"},
-		{Name: "port", Type: filament.FieldInt, Default: defaultPort, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "PostgreSQL server port"},
-		{Name: "username", Type: filament.FieldString, Required: true, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "PostgreSQL username"},
-		{Name: "password", Type: filament.FieldSecret, Required: true, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "PostgreSQL password"},
-		{Name: "database_name", Type: filament.FieldString, Required: true, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "Database to connect to"},
-		{Name: "ssl_mode", Type: filament.FieldEnum, Default: "prefer", Enum: []filament.EnumOption{
-			{Value: "disable", Label: "Disable"},
-			{Value: "allow", Label: "Allow"},
-			{Value: "prefer", Label: "Prefer"},
-			{Value: "require", Label: "Require"},
-			{Value: "verify-ca", Label: "Verify CA"},
-			{Value: "verify-full", Label: "Verify Full"},
-		}, Scope: filament.ScopeConnection, VisibleWhen: fields, Help: "PostgreSQL TLS verification mode"},
-	}
+	DSN          string
+	DriverConfig *pgxpool.Config
 }
 
 // Resolve translates connector configuration into a PostgreSQL pool config.
@@ -62,7 +38,7 @@ func Resolve(cfg filament.Config) (Resolved, error) {
 	if err != nil {
 		return Resolved{}, fmt.Errorf("parse dsn: %w", err)
 	}
-	return Resolved{DSN: dsn, Pool: pool}, nil
+	return Resolved{DSN: dsn, DriverConfig: pool}, nil
 }
 
 func fieldsDSN(cfg filament.Config) (string, error) {
